LayoutTests/platform/chromium/css3/compositing/background-blend-mode-property-parsing-expected.txt

Test the parsing of the -webkit-background-blend-mode property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Basic reference : normal
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be normal (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSPrimitiveValue. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSPrimitiveValuePrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSPrimitiveValueConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : multiply, screen
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be multiply, screen (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSValueList. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSValueListPrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSValueListConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : screen, hue
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be screen, hue (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSValueList. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSValueListPrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSValueListConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : overlay, normal
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be overlay, normal (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSValueList. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSValueListPrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSValueListConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : darken, lighten, normal, luminosity
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be darken, lighten, normal, luminosity (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSValueList. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSValueListPrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSValueListConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : lighten
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be lighten (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSPrimitiveValue. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSPrimitiveValuePrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSPrimitiveValueConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : color-dodge
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be color-dodge (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSPrimitiveValue. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSPrimitiveValuePrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSPrimitiveValueConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : color-burn
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be color-burn (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSPrimitiveValue. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSPrimitiveValuePrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSPrimitiveValueConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : hard-light
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be hard-light (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSPrimitiveValue. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSPrimitiveValuePrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSPrimitiveValueConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : soft-light
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be soft-light (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSPrimitiveValue. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSPrimitiveValuePrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSPrimitiveValueConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : difference
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be difference (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSPrimitiveValue. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSPrimitiveValuePrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSPrimitiveValueConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : exclusion
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be exclusion (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSPrimitiveValue. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSPrimitiveValuePrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSPrimitiveValueConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : hue
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be hue (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSPrimitiveValue. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSPrimitiveValuePrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSPrimitiveValueConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : saturation
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be saturation (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSPrimitiveValue. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSPrimitiveValuePrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSPrimitiveValueConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : color
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be color (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSPrimitiveValue. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSPrimitiveValuePrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSPrimitiveValueConstructor. Threw exception TypeError: Cannot read property 'constructor' of null

Basic reference : luminosity
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-background-blend-mode') should be luminosity (of type string). Was null (of type object).
FAIL jsWrapperClass(blendModeRule) should be CSSPrimitiveValue. Was [null].
FAIL jsWrapperClass(blendModeRule.__proto__) should be CSSPrimitiveValuePrototype. Threw exception TypeError: Cannot read property '__proto__' of null
FAIL jsWrapperClass(blendModeRule.constructor) should be CSSPrimitiveValueConstructor. Threw exception TypeError: Cannot read property 'constructor' of null
PASS successfullyParsed is true

TEST COMPLETE
